Revert "Web Inspector: move Console.addInspectedNode to DOM.setInspectedNode"
To be able to use new remote inspector with WebKitGTK 2.18.

Otherwise, the following exception occurs:

TypeError:  ConsoleAgent.addInspectedNode is not a function. (In 'ConsoleAgent.addInspectedNode(selectedDOMNode.id) ', 'ConsoleAgent.addInspectedNode' is undefined)  (at DOMTreeContentView.js: 434: 42)
    _selectedNodeDidChange @ DOMTreeContentView.js: 434: 42
    dispatch @ Object.js: 170: 30
    dispatchEventToListeners @ Object.js: 177: 17
    _selectedNodeChanged @ DOMTreeOutline.js: 200: 38
    selectDOMNode @ DOMTreeOutline.js: 138: 38
    selectAndRevealDOMNode @ DOMTreeContentView.js: 191: 43
    selectNode @ DOMTreeContentView.js: 388: 40
    _restoreSelectedNodeAfterUpdate @ DOMTreeContentView.js: 407: 28
    _rootDOMNodeAvailable @ FrameDOMTreeContentView.js: 73: 45
    _rootDOMNodeAvailable @ [native code]
    requestRootDOMNode @ DOMTree.js: 94: 21
    _requestRootDOMNode @ FrameDOMTreeContentView.js: 83: 41
    _rootDOMNodeInvalidated @ FrameDOMTreeContentView.js: 78: 33
    dispatch @ Object.js: 170: 30
    dispatchEventToListeners @ Object.js: 177: 17
    performInvalidate @ DOMTree.js: 80: 42
    performInvalidate @ [native code]

This reverts commit ea69cfa024140572cb74f27d7dbf58b96139feaf.
