+++ LayoutTests/ChangeLog
@@ -18,21 +18,6 @@
 
 2017-09-25  Devin Rousso  <webkit@devinrousso.com>
 
-        Web Inspector: move Console.addInspectedNode to DOM.setInspectedNode
-        https://bugs.webkit.org/show_bug.cgi?id=176827
-
-        Reviewed by Joseph Pecoraro.
-
-        * http/tests/inspector/dom/cross-domain-inspected-node-access-expected.txt: Renamed from LayoutTests/http/tests/inspector/console/cross-domain-inspected-node-access-expected.txt.
-        * http/tests/inspector/dom/cross-domain-inspected-node-access.html: Renamed from LayoutTests/http/tests/inspector/console/cross-domain-inspected-node-access.html.
-        * inspector/console/command-line-api.html:
-        * inspector/dom/setInspectedNode-expected.txt: Renamed from LayoutTests/inspector/console/addInspectedNode-expected.txt.
-        * inspector/dom/setInspectedNode.html: Renamed from LayoutTests/inspector/console/addInspectedNode.html.
-        * platform/gtk/TestExpectations:
-        * platform/mac/TestExpectations:
-
-2017-09-25  Devin Rousso  <webkit@devinrousso.com>
-
         Web Inspector: Add autocompletion suggestions for CSS attr based on the selected element's attributes
         https://bugs.webkit.org/show_bug.cgi?id=177346
