+++ LayoutTests/inspector/console/addInspectedNode-expected.txt
@@ -1,7 +1,7 @@
-Test for the DOM.setInspectedNode command.
+Test for the Console.addInspectedNode command.
 
 
-== Running test suite: DOM.setInspectedNode
+== Running test suite: Console.addInspectedNode
 -- Running test case: Check$0IsUndefined
 PASS: $0 should be undefined.
