+++ LayoutTests/platform/gtk/TestExpectations
@@ -2049,6 +2049,7 @@ webkit.org/b/149916 http/tests/inspector/dom/disconnect-dom-tree-after-main-fram
 webkit.org/b/149916 inspector/codemirror/prettyprinting-css.html [ Pass Slow ]
 webkit.org/b/149916 inspector/codemirror/prettyprinting-css-rules.html [ Pass Timeout ]
 webkit.org/b/149916 inspector/codemirror/prettyprinting-javascript.html [ Pass Slow ]
+webkit.org/b/149916 inspector/console/addInspectedNode.html [ Pass Slow ]
 webkit.org/b/149916 inspector/console/clearMessages.html [ Pass Slow ]
 webkit.org/b/149916 inspector/console/command-line-api.html [ Pass Slow ]
 webkit.org/b/149916 inspector/console/console-api.html [ Pass Timeout ]
@@ -2069,7 +2070,6 @@ webkit.org/b/149916 inspector/css/stylesheet-events-imports.html [ Pass Slow ]
 webkit.org/b/149916 inspector/css/stylesheet-events-inspector-stylesheet.html [ Pass Slow ]
 webkit.org/b/149916 inspector/css/stylesheet-events-multiple-documents.html [ Pass Slow ]
 webkit.org/b/149916 inspector/css/stylesheet-with-mutations.html [ Pass Slow ]
-webkit.org/b/149916 inspector/dom/setInspectedNode.html [ Pass Slow ]
 webkit.org/b/149916 inspector/heap/garbageCollected.html [ Pass Slow ]
 webkit.org/b/149916 inspector/heap/gc.html [ Pass Slow ]
 webkit.org/b/149916 inspector/network/client-blocked-load.html [ Pass Slow ]
@@ -2135,7 +2135,7 @@ webkit.org/b/116958 http/tests/navigation/scrollstate-after-location-reload.html
 
 webkit.org/b/120682 inspector/page/archive.html [ Timeout ]
 
-webkit.org/b/122571 http/tests/inspector/dom/cross-domain-inspected-node-access.html [ Timeout Pass ]
+webkit.org/b/122571 http/tests/inspector/console/cross-domain-inspected-node-access.html [ Timeout Pass ]
 webkit.org/b/122571 inspector/debugger/setBreakpoint.html [ Timeout Pass ]
 webkit.org/b/122571 inspector/debugger/call-frame-function-name.html [ Timeout Pass ]
 webkit.org/b/122571 inspector/debugger/setBreakpoint-options-exception.html [ Timeout Pass ]
