+++ Source/JavaScriptCore/inspector/agents/InspectorConsoleAgent.h
@@ -62,6 +62,7 @@ public:
     void disable(ErrorString&) override;
     void clearMessages(ErrorString&) override;
     void setMonitoringXHREnabled(ErrorString&, bool enabled) override = 0;
+    void addInspectedNode(ErrorString&, int nodeId) override = 0;
 
     bool enabled() const { return m_enabled; }
     void reset();
