+++ Source/JavaScriptCore/inspector/agents/JSGlobalObjectConsoleAgent.h
@@ -38,6 +38,7 @@ public:
 
     // FIXME: XHRs and Nodes only makes sense debugging a Web context. Can this be moved to a different agent?
     void setMonitoringXHREnabled(ErrorString&, bool enabled) override;
+    void addInspectedNode(ErrorString&, int nodeId) override;
 };
 
 } // namespace Inspector
