+++ Source/JavaScriptCore/inspector/protocol/Console.json
@@ -64,6 +64,13 @@
             "parameters": [
                 { "name": "enabled", "type": "boolean", "description": "Monitoring enabled state." }
             ]
+        },
+        {
+            "name": "addInspectedNode",
+            "description": "Enables console to refer to the node with given id via $0 (see Command Line API for more details).",
+            "parameters": [
+                { "name": "nodeId", "$ref": "DOM.NodeId", "description": "DOM node id to be accessible by means of $0 command line API." }
+            ]
         }
     ],
     "events": [
