+++ Source/WebCore/inspector/WorkerConsoleAgent.cpp
@@ -35,4 +35,9 @@ WorkerConsoleAgent::WorkerConsoleAgent(WorkerAgentContext& context, InspectorHea
 {
 }
 
+void WorkerConsoleAgent::addInspectedNode(ErrorString& errorString, int)
+{
+    errorString = ASCIILiteral("Unsupported for Workers.");
+}
+
 } // namespace WebCore
