+++ Source/WebCore/inspector/WorkerConsoleAgent.h
@@ -36,6 +36,9 @@ class WorkerConsoleAgent final : public WebConsoleAgent {
 public:
     WorkerConsoleAgent(WorkerAgentContext&, Inspector::InspectorHeapAgent*);
     ~WorkerConsoleAgent() { }
+
+private:
+    void addInspectedNode(ErrorString&, int nodeId) override;
 };
 
 } // namespace WebCore
